refactor: revert namespace change
