+++ src/Umbraco.Community.SimpleWorkspaceViews/Web/NullView.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.AspNetCore.Mvc.ViewEngines;
 
-namespace Umbraco.Community.SimpleWorkspaceViews.Web.Models;
+namespace Umbraco.Community.SimpleWorkspaceViews.Web;
 
 internal sealed class NullView : IView
 {
